Move wallet selectors to a properly-named file
