+++ src/actions/SettingsActions.js
@@ -15,7 +15,7 @@ import s from '../locales/strings.js'
 import * as ACCOUNT_SETTINGS from '../modules/Core/Account/settings.js'
 import { updateExchangeRates } from '../modules/ExchangeRates/action.js'
 import { sendLogs } from '../modules/Logs/action.js'
-import { convertCurrency } from '../modules/UI/selectors.js'
+import { convertCurrency } from '../selectors/WalletSelectors.js'
 import { THEME } from '../theme/variables/airbitz.js'
 import { type Dispatch, type GetState, type RootState } from '../types/reduxTypes.js'
