+++ src/components/modals/FlipInputModal.js
@@ -11,8 +11,8 @@ import { sprintf } from 'sprintf-js'
 import { updateMaxSpend, updateTransactionAmount } from '../../actions/SendConfirmationActions.js'
 import * as Constants from '../../constants/indexConstants'
 import s from '../../locales/strings.js'
-import { convertCurrencyFromExchangeRates, convertNativeToExchangeRateDenomination, getExchangeRate } from '../../modules/UI/selectors.js'
 import { getDisplayDenomination, getExchangeDenomination } from '../../selectors/DenominationSelectors.js'
+import { convertCurrencyFromExchangeRates, convertNativeToExchangeRateDenomination, getExchangeRate } from '../../selectors/WalletSelectors.js'
 import { type RootState } from '../../types/reduxTypes.js'
 import type { GuiCurrencyInfo } from '../../types/types.js'
 import { convertTransactionFeeToDisplayFee, DIVIDE_PRECISION, getDenomFromIsoCode } from '../../util/utils.js'
