+++ src/components/scenes/FioRequestConfirmationScene.js
@@ -12,8 +12,8 @@ import { formatNumber } from '../../locales/intl.js'
 import s from '../../locales/strings.js'
 import { addToFioAddressCache, checkExpiredFioAddress, checkPubAddress } from '../../modules/FioAddress/util'
 import { Slider } from '../../modules/UI/components/Slider/Slider'
-import { getExchangeRate, getSelectedWallet } from '../../modules/UI/selectors.js'
 import { getDisplayDenomination, getPrimaryExchangeDenomination } from '../../selectors/DenominationSelectors.js'
+import { getExchangeRate, getSelectedWallet } from '../../selectors/WalletSelectors.js'
 import { type RootState } from '../../types/reduxTypes'
 import type { GuiCurrencyInfo, GuiDenomination, GuiWallet } from '../../types/types'
 import { emptyCurrencyInfo } from '../../types/types'
