+++ src/components/scenes/FioSentRequestDetailsScene.js
@@ -7,8 +7,8 @@ import { FIAT_CODES_SYMBOLS } from '../../constants/indexConstants'
 import { formatDate, formatNumber } from '../../locales/intl.js'
 import s from '../../locales/strings.js'
 import { isRejectedFioRequest, isSentFioRequest } from '../../modules/FioRequest/util'
-import { getSelectedWallet } from '../../modules/UI/selectors'
 import type { RootState } from '../../reducers/RootReducer'
+import { getSelectedWallet } from '../../selectors/WalletSelectors.js'
 import { type FioRequest, type GuiExchangeRates, type GuiWallet } from '../../types/types.js'
 import { SceneWrapper } from '../common/SceneWrapper'
 import { type Theme, type ThemeProps, cacheStyles, withTheme } from '../services/ThemeContext.js'
