+++ src/components/scenes/NotificationScene.js
@@ -10,7 +10,7 @@ import { connect } from 'react-redux'
 import * as Constants from '../../constants/indexConstants'
 import s from '../../locales/strings'
 import { notif1 } from '../../modules/notifServer.js'
-import { getActiveWalletCurrencyInfos } from '../../modules/UI/selectors'
+import { getActiveWalletCurrencyInfos } from '../../selectors/WalletSelectors.js'
 import { type RootState } from '../../types/reduxTypes.js'
 import { getCurrencyIcon } from '../../util/CurrencyInfoHelpers.js'
 import { SceneWrapper } from '../common/SceneWrapper.js'
